+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Updated dependencies.
+
 ### Added
 
 - Expire inactive notifications after 45 minutes.
+- Added support for Girudo as a game service!
 
 ## [v17.4.0](https://github.com/lexicalunit/spellbot/releases/tag/v17.4.0) - 2026-01-06